Added utilities to get system informatin
