+++ src/zenml/environment.py
@@ -14,10 +14,12 @@
 import os
 import platform
 from importlib.util import find_spec
-from typing import TYPE_CHECKING, Any, Dict, Optional, Tuple, Type, cast
+from pathlib import Path
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Type, cast
 
 import distro
 
+from zenml import __version__
 from zenml.logger import get_logger
 from zenml.utils.singleton import SingletonMetaClass
 
@@ -25,6 +27,53 @@ if TYPE_CHECKING:
     from zenml.steps import StepEnvironment
 
 logger = get_logger(__name__)
+
+
+def get_installed_integrations() -> List[Dict[str, str]]:
+    """Returns list of installed integrations."""
+    from zenml.integrations.registry import integration_registry
+
+    list_of_dicts = []
+    for name, integration_impl in integration_registry.integrations.items():
+        is_installed = integration_impl.check_installation()  # type: ignore[attr-defined]
+        if is_installed:
+            list_of_dicts.append(
+                {
+                    "INSTALLED": "True",
+                    "INTEGRATION": name,
+                }
+            )
+    return list_of_dicts
+
+
+def get_environment() -> str:
+    """Returns a string representing the execution environment of the pipeline.
+    Currently, one of `docker`, `paperspace`, 'colab', or `native`"""
+    if Environment.in_docker():
+        return "docker"
+    elif Environment.in_google_colab():
+        return "colab"
+    elif Environment.in_paperspace_gradient():
+        return "paperspace"
+    elif Environment.in_notebook():
+        return "notebook"
+    else:
+        return "native"
+
+
+def get_system_details():
+    info = {
+        "ZenML version": __version__,
+        "Install Path": Path(__file__).resolve().parent,
+        "Python Version": Environment.python_version(),
+        "Platform Information": Environment.get_system_info(),
+        "Environment": get_environment(),
+        "Integrations": get_installed_integrations(),
+    }
+    return "\n".join(
+        "{:>30} {}".format(k + ":", str(v).replace("\n", " "))
+        for k, v in info.items()
+    )
 
 
 class Environment(metaclass=SingletonMetaClass):
